delete appointment + update appointments mappings

+++ src/main/java/com/example/project/dtos/AppointmentDto.java
@@ -21,11 +21,11 @@ public class AppointmentDto {
     @NotEmpty
     private String type;
     @NotEmpty
-    private int doctorId;
+    private Long doctorId;
     @NotEmpty
     private Date date;
     @NotEmpty
-    private int patientId;
+    private Long patientId;
     @NotEmpty
     private String hospitalName;
 }
